+++ tests/bsm2_ol_test_not_connected.py
@@ -136,7 +136,7 @@ def test_bsm2_ol_not_connected():
     timestep = 15 / 24 / 60  # 15 minutes in days
     endtime = 50  # data_in[-1, 0]
     data_time = data_in[:, 0]
-    simtime = np.arange(0, endtime, timestep)
+    simtime = np.arange(0, endtime, timestep, dtype=float)
 
     y_in = data_in[:, 1:]
     del data_in
