+++ tests/primclar_test.py
@@ -44,7 +44,7 @@ def test_primclar():
 
     timestep = 15 / (60 * 24)
     endtime = 200
-    simtime = np.arange(0, endtime, timestep)
+    simtime = np.arange(0, endtime, timestep, dtype=float)
 
     yp_uf = np.zeros(21)
     yp_of = np.zeros(21)
@@ -139,7 +139,7 @@ def test_primclar_dyn():
     timestep = 15 / 24 / 60  # 15 minutes in days
     endtime = 50  # data_in[-1, 0]
     data_time = data_in[:, 0]
-    simtime = np.arange(0, endtime, timestep)
+    simtime = np.arange(0, endtime, timestep, dtype=float)
     y_in = data_in[:, 1:]
     del data_in
