+++ tests/storage_test.py
@@ -59,7 +59,7 @@ def test_storage():
 
     timestep = 15 / (60 * 24)
     endtime = 200
-    simtime = np.arange(0, endtime, timestep)
+    simtime = np.arange(0, endtime, timestep, dtype=float)
 
     yst_out = np.zeros(21)
     yst_vol = 0
@@ -124,7 +124,7 @@ def test_storage_dyn():
     timestep = 15 / 24 / 60  # 15 minutes in days
     endtime = 50  # data_in[-1, 0]
     data_time = data_in[:, 0]
-    simtime = np.arange(0, endtime, timestep)
+    simtime = np.arange(0, endtime, timestep, dtype=float)
     y_in = data_in[:, 1:]
     del data_in
