+++ tests/thickener_test.py
@@ -49,7 +49,7 @@ def test_thickener():
 
     timestep = 15 / (60 * 24)
     endtime = 200
-    simtime = np.arange(0, endtime, timestep)
+    simtime = np.arange(0, endtime, timestep, dtype=float)
 
     yt_uf = np.zeros(21)
     yt_of = np.zeros(21)
@@ -136,7 +136,7 @@ def test_thickener_dyn():
     timestep = 15 / 24 / 60  # 15 minutes in days
     endtime = 50  # data_in[-1, 0]
     data_time = data_in[:, 0]
-    simtime = np.arange(0, endtime, timestep)
+    simtime = np.arange(0, endtime, timestep, dtype=float)
     y_in = data_in[:, 1:]
     del data_in
